refactor: DB를 Repository로 변경, 인터페이스와 구현체 분리
