+++ src/main/java/repository/MemoryUserRepository.java
@@ -1,4 +1,4 @@
-package db;
+package repository;
 
 
 import lombok.extern.slf4j.Slf4j;
@@ -9,19 +9,22 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 @Slf4j
-public class DataBase {
+public class MemoryUserRepository implements UserRepository {
     private static final ConcurrentMap<String, User> users = new ConcurrentHashMap<>();
 
-    public static void addUser(User user) {
+    @Override
+    public void save(User user) {
         users.put(user.getUserId(), user);
         log.info("ADD USER = " + user);
     }
 
-    public static User findUserById(String userId) {
+    @Override
+    public User findUserById(String userId) {
         return users.get(userId);
     }
 
-    public static Collection<User> findAll() {
+    @Override
+    public Collection<User> findAll() {
         return users.values();
     }
 }
