+++ src/main/java/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package repository;
+
+import model.User;
+
+import java.util.Collection;
+
+public interface UserRepository {
+    void save(User user);
+
+    User findUserById(String userId);
+
+    Collection<User> findAll();
+}
